poster-function/requirements.txt
Flask
Pillow
azure-storage-blob
werkzeug
